StudentName 으로 학생을 찾을 수 없음

+++ src/main/java/com/nhnacademy/student/StudentDeleteServlet.java
@@ -9,10 +9,11 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
+import java.util.Objects;
 
 @Slf4j
 @WebServlet(name = "studentDeleteServlet", urlPatterns = "/student/delete")
-public class studentDeleteServlet extends HttpServlet {
+public class StudentDeleteServlet extends HttpServlet {
 
     private StudentRepository studentRepository;
 
@@ -25,6 +26,11 @@ public class studentDeleteServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         //todo get parameter  : id , id가 존재하지 않을경우 throw new RuntimeException("...")
+        String studentId = req.getParameter("id");
+        if(Objects.isNull(studentId)){
+            throw new RuntimeException("StudentId Not Exist!!");
+        }
+
         studentRepository.deleteById(id);
         //todo /student/list <-- redirect
     }
